--- a/scripts/.cspell.json
+++ b/scripts/.cspell.json
@@ -11,6 +11,7 @@
 	"baggins",
 	"biodigital",
 	"bitcast",
+	"bitemporality",
 	"bitfield",
 	"bitrot",
 	"bonwick",