--- a/src/clients/docs_generate.zig
+++ b/src/clients/docs_generate.zig
@@ -263,13 +263,7 @@ const Generator = struct {
         defer env.deinit();
 
         if (run_setup_tests) {
-            // TODO(Phil): Get dotnet tests working outside of linux in CI.
-            // They do work locally on my machines but I can't get them working in CI.
-            if (!(std.mem.eql(u8, self.language.markdown_name, "cs") and
-                builtin.os.tag != .linux))
-            {
-                try env.appendSlice(&[_][]const u8{ "TEST", "true" });
-            }
+            try env.appendSlice(&[_][]const u8{ "TEST", "true" });
         }
         try run_shell_with_env(
             self.arena,