--- a/src/clients/dotnet/TigerBeetle.Benchmarks/TigerBeetle.Benchmarks.csproj
+++ b/src/clients/dotnet/TigerBeetle.Benchmarks/TigerBeetle.Benchmarks.csproj
@@ -1,9 +1,9 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>net6.0</TargetFramework>
+    <TargetFramework>net7.0</TargetFramework>
     <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-    <UseCurrentRuntimeIdentifier>false</UseCurrentRuntimeIdentifier>
+    <UseCurrentRuntimeIdentifier>true</UseCurrentRuntimeIdentifier>
     <RollForward>LatestMajor</RollForward>
   </PropertyGroup>
   <ItemGroup>