--- a/src/clients/dotnet/TigerBeetle.Tests/TigerBeetle.Tests.csproj
+++ b/src/clients/dotnet/TigerBeetle.Tests/TigerBeetle.Tests.csproj
@@ -1,6 +1,6 @@
 <Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
-    <TargetFrameworks>netcoreapp3.1;net6.0</TargetFrameworks>
+    <TargetFrameworks>netcoreapp3.1;net7.0</TargetFrameworks>
     <LangVersion>10</LangVersion>
     <DebugType>Full</DebugType>
     <IsPackable>false</IsPackable>