--- a/src/clients/dotnet/TigerBeetle/TigerBeetle.csproj
+++ b/src/clients/dotnet/TigerBeetle/TigerBeetle.csproj
@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
   <Import Project="TigerBeetle.props" />
   <PropertyGroup>
-    <TargetFrameworks>netstandard2.1;net6.0</TargetFrameworks>
+    <TargetFrameworks>netstandard2.1;net7.0</TargetFrameworks>
     <Platforms>AnyCPU</Platforms>
     <UseCurrentRuntimeIdentifier>false</UseCurrentRuntimeIdentifier>
     <LangVersion>10</LangVersion>
@@ -14,12 +14,8 @@
     <OS Condition="$([MSBuild]::IsOSPlatform('Windows'))">Windows</OS>
   </PropertyGroup>
   <Target Name="BuildZig" BeforeTargets="DispatchToInnerBuilds">
-    <Exec Command=".\zig\zig build dotnet_client -Doptimize=ReleaseSafe"
-          WorkingDirectory="$(MSBuildProjectDirectory)\..\..\..\.."
-          Condition="'$(OS)'=='Windows'"/>
-    <Exec Command="./zig/zig build dotnet_client -Doptimize=ReleaseSafe"
-          WorkingDirectory="$(MSBuildProjectDirectory)/../../../.."
-          Condition="'$(OS)'!='Windows'"/>
+    <Exec Command=".\zig\zig build dotnet_client -Doptimize=ReleaseSafe" WorkingDirectory="$(MSBuildProjectDirectory)\..\..\..\.." Condition="'$(OS)'=='Windows'" />
+    <Exec Command="./zig/zig build dotnet_client -Doptimize=ReleaseSafe" WorkingDirectory="$(MSBuildProjectDirectory)/../../../.." Condition="'$(OS)'!='Windows'" />
   </Target>
   <ItemGroup>
     <Content Include="runtimes\**\*.so">