--- a/src/clients/java/src/main/java/com/tigerbeetle/CreateAccountResultBatch.java
+++ b/src/clients/java/src/main/java/com/tigerbeetle/CreateAccountResultBatch.java
@@ -7,6 +7,7 @@ package com.tigerbeetle;
 
 import java.nio.ByteBuffer;
 
+
 public final class CreateAccountResultBatch extends Batch {
 
     interface Struct {