## This PR:

- Account balances fields are now `u128`.

- Transfers amounts are `u128`.

- Timeouts are `u32` and expressed as `seconds` instead of `nanoseconds`.

- Three `user_data` fields, as `u128`, `u64` and `u32`.

- `AccountMutable` and `AccountImmutable` are now a single groove `Account`.

- Reordered the `Transfer` and `Account` fields to keep `@alignOf(T) == 16`.

## Direct changes:

**Docs:** All changes have been reflected, and new examples have been added for the `user_data_{128|64|32}` fields.

**Benchmarks, Demos, Samples, Tests, and Repl:** The code has been refactored.

**Clients:** Besides the naming and types refactor, these are specific changes for each client:

- Node client: No extensive changes were needed, as `BigInt` was already being used for `u64` values.

- Java client: The API now defaults all `get` acessors to `byte[]` for IDs and user data, and `BigInteger` for balances and amounts. Additionally, overloaded accessors allow the application to get/set values as a pair of primitive `long`. It allows integrating with third-party numerical libraries without the need to allocate temporary objects like arrays or BigIntegers.

- C# client: .NET 7.0 comes with support for `UInt128`. Users can easily opt-in to use the new `UInt128` or to use a replacement type we distribute for .NET Standard 2.1 (which is supported by other runtimes like .NET Core, Xamarin, Unity, and .NET up to 6.0). Also, there are extension methods for converting to/from `BigInteger`.

- Go client: We expose the type `Uint128` with a set of conversion functions, including `math/big.Int`. Likewise, users have the option to use `big.Int` or any other implementation to handle or format 128-bit values.

## Other changes:

- Increased `block_size` to 128K.
Without this, we can't fit everything in a single block due to the size of each object and `lsm_batch_multiple`:
```
src/lsm/schema.zig:180:51: error: overflow of integer type 'u32' with value '-1048'
        const padding_size = constants.block_size - padding_offset;
```

- Argument `--cache-grid` is now validated for the minimum size.
Currently, it's `--cache-grid=256MB` instead of `128MB` due to `block_size` changes.

- Repl parser now accepts identifiers containing numbers (but not starting with numbers).
Example: `user_data_128` would be rejected by the previous version.

<details>
<summary>Benchmarks</summary>

### Before:

```
❯ ./scripts/benchmark.sh
Formatting replica 0...
Starting replica 0...

Benchmarking...
steps [2/4] zig build-exe benchmark ReleaseSafe native-linux... LLVM Emit Object... info: Benchmark running against { 127.0.0.1:43705 }
info(message_bus): connected to replica 0
1225 batches in 73.57 s
load offered = 1000000 tx/s
load accepted = 135926 tx/s
batch latency p00 = 0 ms
batch latency p10 = 18 ms
batch latency p20 = 22 ms
batch latency p30 = 25 ms
batch latency p40 = 26 ms
batch latency p50 = 28 ms
batch latency p60 = 29 ms
batch latency p70 = 29 ms
batch latency p80 = 30 ms
batch latency p90 = 31 ms
batch latency p100 = 3236 ms
transfer latency p00 = 0 ms
transfer latency p10 = 3435 ms
transfer latency p20 = 7277 ms
transfer latency p30 = 12325 ms
transfer latency p40 = 18875 ms
transfer latency p50 = 25889 ms
transfer latency p60 = 33611 ms
transfer latency p70 = 41591 ms
transfer latency p80 = 49480 ms
transfer latency p90 = 56775 ms
transfer latency p100 = 63583 ms
```

### After:

```
❯ ./scripts/benchmark.sh
Formatting replica 0...
Starting replica 0...

Benchmarking...
info: Benchmark running against { 127.0.0.1:38517 }
info(message_bus): connected to replica 0
1226 batches in 66.92 s
load offered = 1000000 tx/s
load accepted = 149436 tx/s
batch latency p00 = 0 ms
batch latency p10 = 17 ms
batch latency p20 = 19 ms
batch latency p30 = 23 ms
batch latency p40 = 24 ms
batch latency p50 = 25 ms
batch latency p60 = 26 ms
batch latency p70 = 28 ms
batch latency p80 = 30 ms
batch latency p90 = 31 ms
batch latency p100 = 2754 ms
transfer latency p00 = 0 ms
transfer latency p10 = 3181 ms
transfer latency p20 = 6575 ms
transfer latency p30 = 11852 ms
transfer latency p40 = 17253 ms
transfer latency p50 = 23811 ms
transfer latency p60 = 30330 ms
transfer latency p70 = 37195 ms
transfer latency p80 = 43995 ms
transfer latency p90 = 50735 ms
transfer latency p100 = 56930 ms
```

</details>