@cb22 2023-08-23
Interesting benchmark results, too. I remember when testing a pure block size increase, that bumped performance too due to the situation wrt compaction pacing (think the sweet spot was 256K).

Would be interesting to run those again (eg, compare `main` @ 128K).