@batiati 2023-08-24
> Interesting benchmark results, too. I remember when testing a pure block size increase, that bumped performance too due to the situation wrt compaction pacing (think the sweet spot was 256K).

The combination of `block_size = 256k` + `lsm_batch_multiple = 128` gives the best TPS. IMO, mostly because it uses more NVMe bandwidth. I think it may change once the compaction pacing is in place, so it's not worth tuning it yet.

> Would be interesting to run those again (eg, compare `main` @ 128K).

The `main` branch with 128k blocks showed the same ~150k TPS we have measured with these PR changes.
The difference is that now we're writing all the 128 bytes of each account during balance updates (no more split between `AccountMutable` and `AccountImmutable`). I could observe about 2GB more writes during the same benchmark, but I think the NVMe could absorb the extra IO and we still gained by not having to do two lookups for the same object.

```
❯ ./scripts/benchmark.sh
Formatting replica 0...
Starting replica 0...

Benchmarking...
steps [2/4] zig build-exe benchmark ReleaseSafe native-linux... LLVM Emit Object... info: Benchmark running against { 127.0.0.1:33091 }
info(message_bus): connected to replica 0
1224 batches in 67.04 s
load offered = 1000000 tx/s
load accepted = 149174 tx/s
batch latency p00 = 0 ms
batch latency p10 = 19 ms
batch latency p20 = 22 ms
batch latency p30 = 25 ms
batch latency p40 = 27 ms
batch latency p50 = 29 ms
batch latency p60 = 30 ms
batch latency p70 = 31 ms
batch latency p80 = 33 ms
batch latency p90 = 35 ms
batch latency p100 = 2569 ms
transfer latency p00 = 0 ms
transfer latency p10 = 3714 ms
transfer latency p20 = 7157 ms
transfer latency p30 = 12014 ms
transfer latency p40 = 17586 ms
transfer latency p50 = 23668 ms
transfer latency p60 = 29589 ms
transfer latency p70 = 36722 ms
transfer latency p80 = 43803 ms
transfer latency p90 = 50881 ms
transfer latency p100 = 57049 ms
```