@batiati 2023-09-15
Note: Dotnet tests are failing for Macos self-hosted.
Please ignore it for this PR, I suspect it is due to cache issues because the self-hosted runner has no isolation when building .Net 7.0 for this PR and .Net 6.0 for everything else.